Expanded money display amount Added TMs and reorganized stores New cutscene

+++ include/item.h
@@ -9,7 +9,7 @@ struct Item
 {
     u8 name[ITEM_NAME_LENGTH];
     u16 itemId;
-    u16 price;
+    u32 price;
     u8 holdEffect;
     u8 holdEffectParam;
     const u8 *description;
@@ -62,7 +62,7 @@ bool8 AddPyramidBagItem(u16 itemId, u16 count);
 bool8 RemovePyramidBagItem(u16 itemId, u16 count);
 const u8 *ItemId_GetName(u16 itemId);
 u16 ItemId_GetId(u16 itemId);
-u16 ItemId_GetPrice(u16 itemId);
+u32 ItemId_GetPrice(u32 itemId);
 u8 ItemId_GetHoldEffect(u16 itemId);
 u8 ItemId_GetHoldEffectParam(u16 itemId);
 const u8 *ItemId_GetDescription(u16 itemId);
